Take care of the compiler warnings. JIRA: CXX-12888
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99402 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/misc/xmlwrapp/dtd.cpp
@@ -211,7 +211,7 @@ namespace {
         try {
             error_messages *p = static_cast<error_messages*>(v);
             if (p) {
-                int     line = xmlLastError.line;
+                long    line = xmlLastError.line;
                 if (line < 0)
                     line = 0;
                 std::string     filename;
